preliminary changes for num.cr support

+++ tests/numpy/ndarray.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-a = np.array([1,2,3,4,5], dtype='int8')
+a = np.array([1,2,3,4,5], dtype='int32')
 if isinstance(a, np.ndarray):
     print('OK')
 else:
